Implement window events using React hooks
