+++ src/react/images.js
@@ -1,0 +1,4 @@
+import pngs from '../demo/assets/*.png'
+import svgs from '../demo/assets/*.svg'
+
+export default { ...pngs, ...svgs }
